Remove casts that are not needed.

+++ src/Credits.cxx
@@ -462,8 +462,7 @@ static void AddStyledText(GUI::ScintillaWindow &wsci, const char *s, int attr) {
 		buf[i*2] = s[i];
 		buf[i*2 + 1] = static_cast<char>(attr);
 	}
-	wsci.AddStyledText(
-	        static_cast<int>(len*2), &buf[0]);
+	wsci.AddStyledText(len*2, &buf[0]);
 }
 
 static void SetAboutStyle(GUI::ScintillaWindow &wsci, int style, Scintilla::API::Colour fore) {
